Adding the values driverName and url to the configuration file.
This will be for future adding other database support. Furthermore a factory method was created for reading the configuration values and creating the Sqlite database layer.

The configuration for the sqlite setup is set to the default database
configuration by setting the values in the configuration file. There
will be also an error thrown if an unsupported database driver is
configured.