Adding container build
Adding Dockerfile for creating a small image via podman or docker. The Dockerfile contains 2 steps. The builder will build the go image. The second step takes the deployable and creates a very small image with the goauth application.

A script was created which can be called for building goauth by podman. Podman is the prefered tool to build the image. By using the force Flag -D, a docker build can be triggered.